Refactor: remove unused classes, fix imports
