+++ src/main/java/net/kodehawa/mantarobot/commands/InfoCmds.java
@@ -47,7 +47,7 @@ import net.kodehawa.mantarobot.utils.commands.EmoteReference;
 import org.json.JSONObject;
 import redis.clients.jedis.Jedis;
 
-import java.awt.*;
+import java.awt.Color;
 import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
