+++ src/main/java/net/kodehawa/mantarobot/commands/LeaderboardCmd.java
@@ -52,7 +52,6 @@ import org.bson.conversions.Bson;
 import redis.clients.jedis.Jedis;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
