+++ src/main/java/net/kodehawa/mantarobot/commands/MarketCmd.java
@@ -60,7 +60,7 @@ import net.kodehawa.mantarobot.utils.commands.EmoteReference;
 import net.kodehawa.mantarobot.utils.commands.ratelimit.IncreasingRateLimiter;
 import net.kodehawa.mantarobot.utils.commands.ratelimit.RatelimitUtils;
 
-import java.awt.*;
+import java.awt.Color;
 import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
