+++ src/main/java/net/kodehawa/mantarobot/commands/PremiumCmds.java
@@ -36,12 +36,12 @@ import net.kodehawa.mantarobot.core.modules.Module;
 import net.kodehawa.mantarobot.core.modules.commands.base.CommandCategory;
 import net.kodehawa.mantarobot.db.entities.Player;
 import net.kodehawa.mantarobot.db.entities.PremiumKey;
-import net.kodehawa.mantarobot.utils.log.LogUtils;
 import net.kodehawa.mantarobot.utils.APIUtils;
 import net.kodehawa.mantarobot.utils.Utils;
 import net.kodehawa.mantarobot.utils.commands.EmoteReference;
+import net.kodehawa.mantarobot.utils.log.LogUtils;
 
-import java.awt.*;
+import java.awt.Color;
 import java.util.Objects;
 import java.util.stream.Collectors;
