+++ src/main/java/net/kodehawa/mantarobot/db/entities/Inventory.java
@@ -15,7 +15,7 @@
  *
  */
 
-package net.kodehawa.mantarobot.db.entities.helpers;
+package net.kodehawa.mantarobot.db.entities;
 
 import net.kodehawa.mantarobot.commands.currency.item.Item;
 import net.kodehawa.mantarobot.commands.currency.item.ItemHelper;
