+++ src/main/java/net/kodehawa/mantarobot/db/entities/Player.java
@@ -30,16 +30,20 @@ import net.kodehawa.mantarobot.commands.currency.profile.inventory.InventorySort
 import net.kodehawa.mantarobot.data.Config;
 import net.kodehawa.mantarobot.data.MantaroData;
 import net.kodehawa.mantarobot.db.ManagedMongoObject;
-import net.kodehawa.mantarobot.db.entities.helpers.Inventory;
 import org.bson.codecs.pojo.annotations.BsonId;
 import org.bson.codecs.pojo.annotations.BsonIgnore;
 import org.bson.codecs.pojo.annotations.BsonProperty;
 
 import javax.annotation.Nonnull;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
-import static net.kodehawa.mantarobot.db.entities.helpers.Inventory.serialize;
+import static net.kodehawa.mantarobot.db.entities.Inventory.serialize;
 
 public class Player implements ManagedMongoObject {
     @BsonIgnore
