+++ src/main/java/net/kodehawa/mantarobot/db/entities/UserDatabase.java
@@ -33,7 +33,12 @@ import org.bson.codecs.pojo.annotations.BsonIgnore;
 import javax.annotation.Nonnull;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import static java.lang.System.currentTimeMillis;
