--- src/main/java/net/kodehawa/mantarobot/db/entities/helpers/CustomCommandData.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (C) 2016 Kodehawa
- *
- * Mantaro is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * Mantaro is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Mantaro. If not, see http://www.gnu.org/licenses/
- *
- */
-
-package net.kodehawa.mantarobot.db.entities.helpers;
-
-public class CustomCommandData {
-    private String owner = "";
-    private boolean nsfw = false;
-    private boolean locked = false;
-
-    public CustomCommandData() { }
-
-    public String getOwner() {
-        return this.owner;
-    }
-
-    public void setOwner(String owner) {
-        this.owner = owner;
-    }
-
-    public boolean isNsfw() {
-        return this.nsfw;
-    }
-
-    public void setNsfw(boolean nsfw) {
-        this.nsfw = nsfw;
-    }
-
-    public boolean isLocked() {
-        return locked;
-    }
-
-    public void setLocked(boolean locked) {
-        this.locked = locked;
-    }
-}
